renaming scrape_pdf2 > scrape_pdf
